refactor(test): move integration tests to top-level test/ directory
Group all integration tests under test/ alongside a renamed _helpers.ts.
The directory itself signals 'integration', so the .integration suffix
is dropped from filenames. Update deno.jsonc tasks to use directory
paths instead of glob patterns: test:unit targets denops/, and
test:integration targets test/.
