+++ test/tataku.test.ts
@@ -1,7 +1,7 @@
 import { assertEquals, assertStringIncludes } from "jsr:@std/assert@1.0.13";
 import { test } from "jsr:@denops/test@4.0.0";
-import { prepareStreams } from "./tataku.ts";
-import { addRuntimepath, writeFixture } from "./_test_helpers.ts";
+import { prepareStreams } from "../denops/tataku/tataku.ts";
+import { addRuntimepath, writeFixture } from "./_helpers.ts";
 
 const collectorSource = `
 export default () =>
